refactor: move filtering to wasm-storage
