+++ x/wasm-storage/keeper/filter_test.go
@@ -1,11 +1,9 @@
-package drfilters_test
+package keeper_test
 
 import (
-	"github.com/sedaprotocol/seda-chain/drfilters"
 	"testing"
 
 	"github.com/sedaprotocol/seda-chain/x/wasm-storage/keeper"
-
 	"github.com/stretchr/testify/require"
 )
 
@@ -14,26 +12,26 @@ func TestOutliers_None(t *testing.T) {
 		name       string
 		tallyInput []byte
 		want       []bool
-		reveals    []drfilters.Reveller
+		reveals    []keeper.RevealBody
 		wantErr    error
 	}{
 		{
 			name:       "None filter",
 			tallyInput: []byte{193, 128}, // filterProp{ Algo: 0}
 			want:       []bool{false, false, false, false, false},
-			reveals: []drfilters.Reveller{
-				keeper.RevealBody{},
-				keeper.RevealBody{},
-				keeper.RevealBody{},
-				keeper.RevealBody{},
-				keeper.RevealBody{},
+			reveals: []keeper.RevealBody{
+				{},
+				{},
+				{},
+				{},
+				{},
 			},
 			wantErr: nil,
 		},
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			got, _, err := drfilters.Outliers(tt.tallyInput, tt.reveals)
+			got, _, err := keeper.Outliers(tt.tallyInput, tt.reveals)
 			if tt.wantErr != nil {
 				require.ErrorIs(t, err, tt.wantErr)
 				return
